requirements.txt
torch>=2.0
torchvision>=0.15
pillow>=9.0
numpy>=1.22